Deduplicate stacktrace/{attach,poll}_rtt in run/test cmds
