+++ probe-rs/src/bin/probe-rs/cmd/test.rs
@@ -1,5 +1,4 @@
 use std::cell::RefCell;
-use std::io::Write;
 use std::ops::Range;
 use std::path::Path;
 use std::sync::atomic::{AtomicBool, Ordering};
@@ -8,10 +7,9 @@ use std::time::{Duration, Instant};
 
 use anyhow::{anyhow, bail, Result};
 use libtest_mimic::{Failed, Trial};
-use probe_rs::debug::{DebugInfo, DebugRegisters};
 use probe_rs::{
-    exception_handler_for_core, BreakpointCause, Core, CoreInterface, CoreStatus, Error,
-    HaltReason, Lister, MemoryInterface, SemihostingCommand, Session, VectorCatchCondition,
+    BreakpointCause, Core, CoreStatus, Error, HaltReason, Lister, MemoryInterface,
+    SemihostingCommand, Session, VectorCatchCondition,
 };
 use probe_rs_target::MemoryRegion;
 use signal_hook::consts::signal;
@@ -20,10 +18,9 @@ use time::UtcOffset;
 
 use crate::util::common_options::{BinaryDownloadOptions, ProbeOptions};
 use crate::util::flash::{build_loader, run_flash_download};
-use crate::util::rtt::{self, RttConfig};
+use crate::util::rtt::{self, poll_rtt, try_attach_to_rtt};
+use crate::util::stack_trace::print_stacktrace;
 use crate::FormatOptions;
-
-const RTT_RETRIES: usize = 10;
 
 #[derive(clap::Parser)]
 pub struct Cmd {
@@ -383,7 +380,7 @@ fn run_loop(
     });
     let path = Path::new(path);
 
-    let mut rtta = attach_to_rtt(
+    let mut rtta = try_attach_to_rtt(
         core,
         memory_map,
         rtt_scan_regions,
@@ -510,112 +507,4 @@ fn run_loop(
     signal_hook::flag::register_conditional_default(signal::SIGINT, exit)?;
 
     result
-}
-
-/// Prints the stacktrace of the current execution state.
-fn print_stacktrace(core: &mut impl CoreInterface, path: &Path) -> Result<(), anyhow::Error> {
-    let Some(debug_info) = DebugInfo::from_file(path).ok() else {
-        log::error!("No debug info found.");
-        return Ok(());
-    };
-    let initial_registers = DebugRegisters::from_core(core);
-    let exception_interface = exception_handler_for_core(core.core_type());
-    let instruction_set = core.instruction_set().ok();
-    let stack_frames = debug_info
-        .unwind(
-            core,
-            initial_registers,
-            exception_interface.as_ref(),
-            instruction_set,
-        )
-        .unwrap();
-    for (i, frame) in stack_frames.iter().enumerate() {
-        print!("Frame {}: {} @ {}", i, frame.function_name, frame.pc);
-
-        if frame.is_inlined {
-            print!(" inline");
-        }
-        println!();
-
-        if let Some(location) = &frame.source_location {
-            if location.directory.is_some() || location.file.is_some() {
-                print!("       ");
-
-                if let Some(dir) = &location.directory {
-                    print!("{}", dir.to_path().display());
-                }
-
-                if let Some(file) = &location.file {
-                    print!("/{file}");
-
-                    if let Some(line) = location.line {
-                        print!(":{line}");
-
-                        if let Some(col) = location.column {
-                            match col {
-                                probe_rs::debug::ColumnType::LeftEdge => {
-                                    print!(":1")
-                                }
-                                probe_rs::debug::ColumnType::Column(c) => {
-                                    print!(":{c}")
-                                }
-                            }
-                        }
-                    }
-                }
-
-                println!();
-            }
-        }
-    }
-    Ok(())
-}
-
-/// Poll RTT and print the received buffer.
-fn poll_rtt(
-    rtta: &mut Option<rtt::RttActiveTarget>,
-    core: &mut Core<'_>,
-    stderr: &mut std::io::Stderr,
-) -> Result<bool, anyhow::Error> {
-    let mut had_data = false;
-    if let Some(rtta) = rtta {
-        for (_ch, data) in rtta.poll_rtt_fallible(core)? {
-            if !data.is_empty() {
-                had_data = true;
-            }
-            stderr.write_all(data.as_bytes())?;
-        }
-    };
-    Ok(had_data)
-}
-
-/// Attach to the RTT buffers.
-fn attach_to_rtt(
-    core: &mut Core<'_>,
-    memory_map: &[MemoryRegion],
-    scan_regions: &[Range<u64>],
-    path: &Path,
-    rtt_config: RttConfig,
-    timestamp_offset: UtcOffset,
-    log_format: Option<&str>,
-) -> Option<rtt::RttActiveTarget> {
-    for _ in 0..RTT_RETRIES {
-        match rtt::attach_to_rtt(
-            core,
-            memory_map,
-            scan_regions,
-            path,
-            &rtt_config,
-            timestamp_offset,
-            log_format,
-        ) {
-            Ok(target_rtt) => return Some(target_rtt),
-            Err(error) => {
-                log::debug!("{:?} RTT attach error", error);
-            }
-        }
-        std::thread::sleep(std::time::Duration::from_millis(100));
-    }
-    log::error!("Failed to attach to RTT continuing...");
-    None
 }
