+++ probe-rs/src/bin/probe-rs/util/stack_trace.rs
@@ -1,0 +1,65 @@
+use std::path::Path;
+
+use probe_rs::{
+    debug::{DebugInfo, DebugRegisters},
+    exception_handler_for_core, CoreInterface,
+};
+
+/// Prints the stacktrace of the current execution state.
+pub fn print_stacktrace(core: &mut impl CoreInterface, path: &Path) -> Result<(), anyhow::Error> {
+    let Some(debug_info) = DebugInfo::from_file(path).ok() else {
+        log::error!("No debug info found.");
+        return Ok(());
+    };
+    let initial_registers = DebugRegisters::from_core(core);
+    let exception_interface = exception_handler_for_core(core.core_type());
+    let instruction_set = core.instruction_set().ok();
+    let stack_frames = debug_info
+        .unwind(
+            core,
+            initial_registers,
+            exception_interface.as_ref(),
+            instruction_set,
+        )
+        .unwrap();
+    for (i, frame) in stack_frames.iter().enumerate() {
+        print!("Frame {}: {} @ {}", i, frame.function_name, frame.pc);
+
+        if frame.is_inlined {
+            print!(" inline");
+        }
+        println!();
+
+        if let Some(location) = &frame.source_location {
+            if location.directory.is_some() || location.file.is_some() {
+                print!("       ");
+
+                if let Some(dir) = &location.directory {
+                    print!("{}", dir.to_path().display());
+                }
+
+                if let Some(file) = &location.file {
+                    print!("/{file}");
+
+                    if let Some(line) = location.line {
+                        print!(":{line}");
+
+                        if let Some(col) = location.column {
+                            match col {
+                                probe_rs::debug::ColumnType::LeftEdge => {
+                                    print!(":1")
+                                }
+                                probe_rs::debug::ColumnType::Column(c) => {
+                                    print!(":{c}")
+                                }
+                            }
+                        }
+                    }
+                }
+
+                println!();
+            }
+        }
+    }
+    Ok(())
+}
